# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: API definition for Umiverse server apis
  title: Umiverse Server API Spec
  version: 2.0.0
servers:
- url: https://api-dev.umiverse.io/
...
paths:
...
  /wallet/balance:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceInfo'
          description: query user balance success.
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: balance info not found.
        "500":
          description: balance info query internal error.
      security:
      - bearerAuth: []
      summary: 当前用户账户总额
      tags:
      - wallet
      x-accepts: application/json
      x-tags:
      - tag: wallet
...
components:
  schemas:
...
    BalanceInfo:
      description: user balance info
      example:
        value: 0
        account: account
      properties:
        account:
          description: account of user web3Auth.
          type: string
        value:
          description: quantity of user account.
          format: int64
          type: integer
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http